bloglist done up to 4.23

+++ osa4/blogilista/controllers/blogs.js
@@ -1,6 +1,6 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
-const User = require('../models/user');
+const { userExtractor } = require('../utils/middleware');
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
@@ -17,16 +17,25 @@ blogsRouter.get('/:id', (request, response) => {
   });
 });
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
+  const user = request.user;
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) {
+    return response.status(404).json({ error: 'Blog not found' });
+  }
+  if (user._id.toString() !== blog.user.toString()) {
+    return response.status(403).end();
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
 
   response.status(204).end();
 });
 
-blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body;
-
-  const user = await User.findById(body.userId);
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const { body, user } = request;
 
   const blog = new Blog({ ...body, user: user._id });
 
@@ -35,7 +44,7 @@ blogsRouter.post('/', async (request, response, next) => {
   user.blogs = user.blogs.concat(savedBlog._id);
   await user.save();
 
-  response.json(savedBlog);
+  response.status(201).json(savedBlog);
 });
 
 blogsRouter.put('/:id', async (request, response, next) => {
